query/requirements.txt
sentence-transformers==2.2.1
psycopg2-binary==2.9.1
pgvector
utils
accelerate